requirements.txt
aiogram>=3.4,<4
pydantic>=2.6,<3
python-dotenv>=1.0,<2
aiohttp>=3.9,<4